c: Snackbar moved to 'ui' folder
