+++ client/src/components/AppLayout/index.tsx
@@ -2,7 +2,7 @@ import React, { useState, useEffect, useRef } from 'react'
 import styled from 'styled-components'
 import Nav from '@/components/Nav'
 import Player from '@/components/Player'
-import Snackbar from '@/components/Snackbar'
+import Snackbar from '@/components/ui/Snackbar'
 import MiniPlayer from '@/components/MiniPlayer'
 import AudioPlayer from '@/player'
 import { timeFormater } from '@/utils/helpers'
